requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
requests
django-cors-middleware
python-dotenv
django-safedelete
djangorestframework-simplejwt
django-auditlog==2.0.0
websocket-client
websockets
django-rest-auth==0.9.5
cryptography
PyYAML==5.4.1
python-dateutil
gunicorn
django-auditlog